requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
mlflow==2.22.4
numpy==2.0.2
pandas==2.2.3
scikit-learn==1.8.0

# Monitoring
prometheus-client==0.21.1
evidently==0.6.7
datasets==3.6.0

# Tests
pytest==8.3.5
pytest-asyncio==0.24.0
httpx==0.28.1
python-dotenv==1.0.1
